src/requirements.txt
requests>=2.31.0
playwright>=1.40.0
python-dotenv>=1.0.0
zai-sdk>=0.1.0
sniffio>=1.0.0
beautifulsoup4>=4.14.3